// types/services/firestore.d.ts
export interface Category {
  id: string;
  userId?: string;
  isCustom: boolean;
  name: string;
  type: "expense" | "income";
}
export interface CombinedCategoriesResult {
  categories: Category[];
  lastStandardDoc: QueryDocumentSnapshot<DocumentData> | null;
  lastUserDoc: QueryDocumentSnapshot<DocumentData> | null;
}